Refactor: flytte 'Endring i sykefraværsstatistikken' til felleskomponent
