+++ src/Forside/barnehage/SammenligningIngress/SammenligningIngress.tsx
@@ -1,7 +1,7 @@
 import React, { FunctionComponent } from 'react';
 import { Normaltekst, Systemtittel } from 'nav-frontend-typografi';
 import './SammenligningIngress.less';
-import EndringISykefravRsstatistikkenInfotekst from './EndringISykefraværsstatistikkenInfotekst/EndringISykefraværsstatistikkenInfotekst';
+import EndringISykefravRsstatistikkenInfotekst from '../../../felleskomponenter/EndringISykefraværsstatistikkenInfotekst/EndringISykefraværsstatistikkenInfotekst';
 import { Bransjetype } from '../../../api/virksomhetMetadata';
 
 export const SammenligningIngress: FunctionComponent<{
